Sort threads starting from a base label
The goal is to have the emails to never hit the inbox as that generates a lot of push notification and wearing an Apple Watch that gets super annoying.

So I set up a filter to move all email matching `from:notifications@github.com OR from:noreply@github.com` to skip the inbox (archive) and apply the label "GitHub". GitHub is being used as the base label; any new GH emails are sorted into that folder. Then every couple minutes the script comes along and applies the correct labels depending on the GH repo they are for. On top of that the "GitHub" base label is removed.